requirements.txt
crewai==0.100.1
litellm==1.59.8
langchain-google-genai==2.0.9
google-generativeai==0.8.3
pydantic==2.10.4
pysqlite3-binary
python-dotenv
pyyaml
requests